Added new Functionality
Added getting secret key, api key, buy/sell conditions, and stock symbol. Also added time when bot was added to table. All of this is subject to change, the method names will remain the same but how it will work will change with an updated version of the adapter that i am writing